Enum dslab_mp::logger::LogEntry

pub enum LogEntry {

    NodeStarted {
        time: f64,
        node: String,
        node_id: Id,
    },
    ProcessStarted {
        time: f64,
        node: String,
        proc: String,
    },
    LocalMessageSent {
        time: f64,
        msg_id: String,
        node: String,
        proc: String,
        msg: Message,
    },
    LocalMessageReceived {
        time: f64,
        msg_id: String,
        node: String,
        proc: String,
        msg: Message,
    },
    MessageSent {
        time: f64,
        msg_id: String,
        src_node: String,
        src_proc: String,
        dst_node: String,
        dst_proc: String,
        msg: Message,
    },
    MessageReceived {
        time: f64,
        msg_id: String,
        src_node: String,
        src_proc: String,
        dst_node: String,
        dst_proc: String,
        msg: Message,
    },
    MessageDropped {
        time: f64,
        msg_id: String,
        src_node: String,
        src_proc: String,
        dst_node: String,
        dst_proc: String,
        msg: Message,
    },
    NodeDisconnected {
        time: f64,
        node: String,
    },
    NodeConnected {
        time: f64,
        node: String,
    },
    NodeCrashed {
        time: f64,
        node: String,
    },
    NodeRecovered {
        time: f64,
        node: String,
    },
    TimerSet {
        time: f64,
        timer_id: String,
        timer_name: String,
        node: String,
        proc: String,
        delay: f64,
    },
    TimerFired {
        time: f64,
        timer_id: String,
        timer_name: String,
        node: String,
        proc: String,
    },
    TimerCancelled {
        time: f64,
        timer_id: String,
        timer_name: String,
        node: String,
        proc: String,
    },
    LinkDisabled {
        time: f64,
        from: String,
        to: String,
    },
    LinkEnabled {
        time: f64,
        from: String,
        to: String,
    },
    DropIncoming {
        time: f64,
        node: String,
    },
    PassIncoming {
        time: f64,
        node: String,
    },
    DropOutgoing {
        time: f64,
        node: String,
    },
    PassOutgoing {
        time: f64,
        node: String,
    },
    NetworkPartition {
        time: f64,
        group1: Vec<String>,
        group2: Vec<String>,
    },
    NetworkReset {
        time: f64,
    },
    ProcessStateUpdated {
        time: f64,
        node: String,
        proc: String,
        state: String,
    },
    McStarted {},
    McLocalMessageSent {
        msg: Message,
        proc: String,
    },
    McLocalMessageReceived {
        msg: Message,
        proc: String,
    },
    McMessageSent {
        msg: Message,
        src: String,
        dst: String,
    },
    McMessageReceived {
        msg: Message,
        src: String,
        dst: String,
    },
    McMessageDropped {
        msg: Message,
        src: String,
        dst: String,
    },
    McMessageCorrupted {
        msg: Message,
        corrupted_msg: Message,
        src: String,
        dst: String,
    },
    McMessageDuplicated {
        msg: Message,
        src: String,
        dst: String,
    },
    McTimerSet {
        proc: String,
        timer: String,
    },
    McTimerFired {
        proc: String,
        timer: String,
    },
    McTimerCancelled {
        proc: String,
        timer: String,
    },
    McNodeCrashed {
        node: String,
    },
    McNetworkReset {},
    McNetworkPartition {
        group1: Vec<String>,
        group2: Vec<String>,
    },

}

Represents a logged event.

Variants

NodeStarted

Fields

time: f64

node: String

node_id: Id

ProcessStarted

Fields

time: f64

node: String

proc: String

LocalMessageSent

Fields

time: f64

msg_id: String

node: String

proc: String

msg: Message

LocalMessageReceived

Fields

time: f64

msg_id: String

node: String

proc: String

msg: Message

MessageSent

Fields

time: f64

msg_id: String

src_node: String

src_proc: String

dst_node: String

dst_proc: String

msg: Message

MessageReceived

Fields

time: f64

msg_id: String

src_node: String

src_proc: String

dst_node: String

dst_proc: String

msg: Message

MessageDropped

Fields

time: f64

msg_id: String

src_node: String

src_proc: String

dst_node: String

dst_proc: String

msg: Message

NodeDisconnected

Fields

time: f64

node: String

NodeConnected

Fields

time: f64

node: String

NodeCrashed

Fields

time: f64

node: String

NodeRecovered

Fields

time: f64

node: String

TimerSet

Fields

time: f64

timer_id: String

timer_name: String

node: String

proc: String

delay: f64

TimerFired

Fields

time: f64

timer_id: String

timer_name: String

node: String

proc: String

TimerCancelled

Fields

time: f64

timer_id: String

timer_name: String

node: String

proc: String

LinkDisabled

Link between a pair of nodes is disabled.

Fields

time: f64

from: String

to: String

LinkEnabled

Link between a pair of nodes is enabled.

Fields

time: f64

from: String

to: String

DropIncoming

Dropping of incoming messages for a node is enabled.

Fields

time: f64

node: String

PassIncoming

Dropping of incoming messages for a node is disabled.

Fields

time: f64

node: String

DropOutgoing

Dropping of outgoing messages for a node is enabled.

Fields

time: f64

node: String

PassOutgoing

Dropping of outgoing messages for a node is enabled.

Fields

time: f64

node: String

NetworkPartition

Network partition is occurred between two groups of nodes.

Fields

time: f64

group1: Vec<String>

First group of nodes.

group2: Vec<String>

Second group of nodes.

NetworkReset

Network is reset to normal state (all links are working).

Fields

time: f64

ProcessStateUpdated

Process state is updated.

Fields

time: f64

node: String

proc: String

state: String

String representation of process state.

McStarted

Model checking session is started.

McLocalMessageSent

Fields

msg: Message

proc: String

McLocalMessageReceived

Fields

msg: Message

proc: String

McMessageSent

Fields

msg: Message

src: String

dst: String

McMessageReceived

Fields

msg: Message

src: String

dst: String

McMessageDropped

Fields

msg: Message

src: String

dst: String

McMessageCorrupted

Fields

msg: Message

corrupted_msg: Message

src: String

dst: String

McMessageDuplicated

Fields

msg: Message

src: String

dst: String

McTimerSet

Fields

proc: String

timer: String

McTimerFired

Fields

proc: String

timer: String

McTimerCancelled

Fields

proc: String

timer: String

McNodeCrashed

Fields

node: String

McNetworkReset

McNetworkPartition

Fields

group1: Vec<String>

group2: Vec<String>

Implementations

impl LogEntry

pub fn print(&self)

Prints log entry to console.

pub fn is_mc_message_dropped(&self) -> bool

Checks if event is LogEntry::McMessageDropped.

pub fn is_mc_message_duplicated(&self) -> bool

Checks if event is LogEntry::McMessageDuplicated.

pub fn is_mc_message_sent(&self) -> bool

Checks if event is LogEntry::McMessageSent.

pub fn is_mc_local_message_sent(&self) -> bool

Checks if event is LogEntry::McLocalMessageSent.

pub fn is_mc_message_received(&self) -> bool

Checks if event is LogEntry::McLocalMessageReceived.

pub fn is_mc_timer_set(&self) -> bool

Checks if event is LogEntry::McTimerSet.

pub fn is_mc_timer_fired(&self) -> bool

Checks if event is LogEntry::McTimerFired.