Enum dslab_mp::mc::events::McEvent

pub enum McEvent {
    MessageReceived {
        msg: Message,
        src: String,
        dst: String,
        options: DeliveryOptions,
    },
    TimerFired {
        proc: String,
        timer: String,
        timer_delay: McTime,
    },
    TimerCancelled {
        proc: String,
        timer: String,
    },
    MessageDropped {
        msg: Message,
        src: String,
        dst: String,
        receive_event_id: Option<McEventId>,
    },
    MessageDuplicated {
        msg: Message,
        src: String,
        dst: String,
        receive_event_id: McEventId,
    },
    MessageCorrupted {
        msg: Message,
        corrupted_msg: Message,
        src: String,
        dst: String,
        receive_event_id: McEventId,
    },
}

Special events used in model checking instead of standard events.

Variants

MessageReceived

Fields

msg: Message

The message itself.

src: String

The name of the process that sent the message.

dst: String

The name of the process that received the message.

options: DeliveryOptions

Network delivery options for the message.

The event of receiving a non-local message by a process.

TimerFired

Fields

proc: String

The process to which the timer belongs to.

timer: String

The timer name.

timer_delay: McTime

The timer duration.

The event of a timer expiration.

TimerCancelled

Fields

proc: String

The process to which the timer belongs to.

timer: String

The timer name.

The event of cancelling a timer.

MessageDropped

Fields

msg: Message

The dropped message itself.

src: String

The name of the process that sent the message.

dst: String

The name of the process the message was sent to.

receive_event_id: Option<McEventId>

The id of original MessageReceived event.

Can be None if the message is dropped unconditionally.

The event of dropping a message. Created by the model checking strategy.

MessageDuplicated

Fields

msg: Message

The duplicated message itself.

src: String

The name of the process that sent the message.

dst: String

The name of the process the message was sent to.

receive_event_id: McEventId

The id of original MessageReceived event.

The event of duplicating a message. Created by model checking strategy.

MessageCorrupted

Fields

msg: Message

The original message.

corrupted_msg: Message

The message after corruption.

src: String

The name of the process that sent the message.

dst: String

The name of the process the message was sent to.

receive_event_id: McEventId

The id of original MessageReceived event.

The event of corrupting a message. Created by model checking strategy.

Implementations

impl McEvent

pub fn duplicate(&self) -> Option<McEvent>

Duplicates the MessageReceived event with decreased max_dupl_count.

pub fn disable_duplications(&mut self)

Modifies event so it won’t be duplicated in future.

pub fn to_log_entry(&self) -> LogEntry

Converts McEvent to LogEntry.