Module dslab_mp::mc::predicates::invariants

Invariants check whether state is correct or not.

Functions

all_invariants

Combines multiple invariant functions by returning Ok iff all invariants are satisfied.

received_messages

Verifies that the set of local messages delivered by a process matches exactly the expected messages. Message duplications or unexpected messages are not allowed.

state_depth

Checks that state depth does not exceed the given value.

state_depth_current_run

Checks that state depth for current run does not exceed the given value.

time_limit

Checks that overall run duration does not exceed the given time limit.