Module dslab_mp::mc::predicates::prunes

Prunes cut execution branches if further analysis is considered unnecessary or computation-heavy.

Functions

any_prune

Combines multiple prune functions by returning Some() iff at least one prune is satisfied.

event_happened_n_times_current_run

Prunes state with at least n events matching the predicate.

events_limit

Prunes states where the number of events matching the predicate is more than the limit.

events_limit_per_proc

Prunes states where the number of events matching the predicate is more than the limit for any of processes.

proc_permutations

Prunes states where processes are mentioned in any permutation except the given one. It helps to create symmetry-breaking constraints

sent_messages_limit

Prunes states where some process has sent more messages than the given value.

state_depth

Prunes states with depth exceeding the given value.