Struct dslab_mp::network::Network

pub struct Network { /* private fields */ }

Represents a network that transmits messages between processes located on different nodes.

Implementations

impl Network

pub fn add_node(&mut self, name: String, id: Id)

Adds a new node to the network.

pub fn proc_locations(&self) -> &HashMap<String, String>

Returns a map with process locations (process -> node).

pub fn set_proc_location(&mut self, proc: String, node: String)

Sets the process location.

pub fn max_delay(&self) -> f64

Returns the maximum network delay.

pub fn set_delay(&mut self, delay: f64)

Sets the fixed network delay.

pub fn set_delays(&mut self, min_delay: f64, max_delay: f64)

Sets the minimum and maximum network delays.

pub fn drop_rate(&self) -> f64

Returns the message drop probability.

pub fn set_drop_rate(&mut self, drop_rate: f64)

Sets the message drop probability.

pub fn dupl_rate(&self) -> f64

Returns the message duplication probability.

pub fn set_dupl_rate(&mut self, dupl_rate: f64)

Sets the message duplication probability.

pub fn corrupt_rate(&self) -> f64

Returns the message corruption probability.

pub fn set_corrupt_rate(&mut self, corrupt_rate: f64)

Sets the message corruption probability.

pub fn get_drop_incoming(&self) -> &HashSet<String>

Returns nodes with enabled dropping of incoming messages.

pub fn drop_incoming(&mut self, node: &str)

Enables dropping of incoming messages for a node.

pub fn pass_incoming(&mut self, node: &str)

Disables dropping of incoming messages for a node.

pub fn get_drop_outgoing(&self) -> &HashSet<String>

Returns nodes with enabled dropping of outgoing messages.

pub fn drop_outgoing(&mut self, node: &str)

Enables dropping of outgoing messages for a node.

pub fn pass_outgoing(&mut self, node: &str)

Disables dropping of outgoing messages for a node.

pub fn disconnect_node(&mut self, node: &str)

Disconnects a node from the network.

Equivalent to enabling dropping of both incoming and outgoing messages for a node.

pub fn connect_node(&mut self, node: &str)

Connects a node to a network.

Equivalent to disabling dropping of both incoming and outgoing messages for a node.

pub fn disabled_links(&self) -> &HashSet<(String, String)>

Returns disabled links.

pub fn disable_link(&mut self, from: &str, to: &str)

Disables link between nodes from and to by dropping all messages sent in this direction.

pub fn enable_link(&mut self, from: &str, to: &str)

Enables link between nodes from and to.

pub fn make_partition(&mut self, group1: &[&str], group2: &[&str])

Creates a network partition between two groups of nodes.

pub fn reset(&mut self)

Resets the network links by enabling all links and disabling dropping of incoming/outgoing messages for all nodes.

Note that this does not affect the drop_rate setting.

pub fn network_message_count(&self) -> u64

Returns the total number of messages sent via the network.

pub fn traffic(&self) -> u64

Returns the total size of messages sent via the network.