Trait dslab_mp::process::ProcessState

pub trait ProcessState: Downcast + Debug {
    // Required methods
    fn hash_with_dyn(&self, hasher: &mut dyn Hasher);
    fn eq_with_dyn(&self, other: &dyn ProcessState) -> bool;
}

A trait for implementations of process state.

Required Methods

fn hash_with_dyn(&self, hasher: &mut dyn Hasher)

Computes a hash of process state using the passed hasher.

fn eq_with_dyn(&self, other: &dyn ProcessState) -> bool

Tests for self and other values to be equal.

Implementations

impl dyn ProcessState

pub fn is<__T: ProcessState>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: ProcessState>( self: Box<Self> ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: ProcessState>( self: Rc<Self> ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: ProcessState>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: ProcessState>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.