Struct dslab_mp::system::System

pub struct System { /* private fields */ }

Models distributed system consisting of multiple nodes connected via network.

Implementations

impl System

pub fn new(seed: u64) -> Self

Creates a system with specified random seed.

pub fn with_log_file(seed: u64, log_path: &Path) -> Self

Creates a system with logging events to file.

pub fn logger(&self) -> RefMut<'_, Logger>

Returns a mutable reference to system logger.

pub fn network(&self) -> RefMut<'_, Network>

Returns a mutable reference to network.

pub fn nodes(&self) -> Vec<String>

Returns a list of node names.

pub fn add_node(&mut self, name: &str)

Adds a node to the system.

Note that node names must be unique.

pub fn set_node_clock_skew(&mut self, node: &str, clock_skew: f64)

Sets local clock skew of the node.

pub fn crash_node(&mut self, node_name: &str)

Crashes the specified node.

All pending events created by the node will be discarded. The undelivered messages sent by the node will be dropped. All pending and future events destined to the node will be discarded.

Processes running on the node are not cleared to allow working with processes after the crash (i.e. examine event log).

pub fn recover_node(&mut self, node_name: &str)

Recovers the previously crashed node.

Processes running on the node before the crash are cleared. The delivery of events to the node is enabled.

pub fn get_node(&self, name: &str) -> Option<Ref<'_, Node>>

Returns an immutable reference to the node.

pub fn get_mut_node(&self, name: &str) -> Option<RefMut<'_, Node>>

Returns a mutable reference to the node.

pub fn node_is_crashed(&self, node: &str) -> bool

Checks if the node is crashed.

pub fn add_process(&mut self, name: &str, proc: Box<dyn Process>, node: &str)

Adds a process executing on the node.

Note that process names should be globally unique.

pub fn process_names(&self) -> Vec<String>

Returns the names of all processes in the system.

pub fn send_local_message(&mut self, proc: &str, msg: Message)

Sends a local message to the process.

pub fn read_local_messages(&mut self, proc: &str) -> Vec<Message>

Reads and returns the local messages produced by the process.

pub fn local_outbox(&self, proc: &str) -> Vec<Message>

Returns a copy of the local messages produced by the process.

In contrast to Self::read_local_messages, this method does not drain the process outbox.

pub fn event_log(&self, proc: &str) -> Vec<EventLogEntry>

Returns the event log for the process.

pub fn max_size(&mut self, proc: &str) -> u64

Returns the maximum size of process inner data observed so far.

pub fn sent_message_count(&self, proc: &str) -> u64

Returns the number of messages sent by the process.

pub fn received_message_count(&self, proc: &str) -> u64

Returns the number of messages received by the process.

pub fn proc_node_name(&self, proc: &str) -> String

Returns the name of node hosting the process.

pub fn proc_node_is_crashed(&self, proc: &str) -> bool

Checks if the node hosting the process is crashed.

pub fn sim(&self) -> &Simulation

Returns the reference to inner simulation.

pub fn time(&self) -> f64

Returns the current simulation time.

pub fn step(&mut self) -> bool

Performs a single step through the simulation.

pub fn steps(&mut self, step_count: u64) -> bool

Performs the specified number of steps through the simulation.

pub fn step_until_no_events(&mut self)

Steps through the simulation until there are no pending events left.

pub fn step_for_duration(&mut self, duration: f64) -> bool

Steps through the simulation with duration limit.

pub fn step_until_local_message( &mut self, proc: &str ) -> Result<Vec<Message>, &str>

Steps through the simulation until the process produces local message(s) or there are no pending events left.

Returns the read local messages if any and error otherwise.

pub fn step_until_local_message_max_steps( &mut self, proc: &str, max_steps: u32 ) -> Result<Vec<Message>, &str>

Similar to Self::step_until_local_message but with additional limit on the number of steps through the simulations.

pub fn step_until_local_message_timeout( &mut self, proc: &str, timeout: f64 ) -> Result<Vec<Message>, &str>

Similar to Self::step_until_local_message but with additional limit on the duration of waiting for local messages.

pub fn gen_range<T, R>(&mut self, range: R) -> T
where T: SampleUniform, R: SampleRange<T>,

Returns a random number in the specified range using the simulation-wide random number generator.

pub fn random_string(&mut self, len: usize) -> String

Returns a random alphanumeric string of specified length using the simulation-wide random number generator.