Module dslab_mp::mc::predicates::collects

Collects select states that should be collected for complex pipelining in MC.

Functions

all_collects

Combines multiple collect functions by returning true iff all collects are satisfied.

any_collect

Combines multiple collect functions by returning true iff at least one collect is satisfied.

event_happened_n_times_current_run

Checks if trace to given state has at least n events matching the predicate.

events_limit

Collects states where the number of events matching the predicate is more than the limit.

got_n_local_messages

Checks if the given process produced n local messages.

no_events

Checks if current state has no more active events.

state_depth

Checks if current state’s depth exceeds the given value.