Trait dslab_mp::mc::strategy::Strategy

pub trait Strategy {

    // Required methods
    fn build(config: StrategyConfig) -> Self
       where Self: Sized;
    fn run(&mut self, system: &mut McSystem) -> McResult;
    fn search_step_impl(
        &mut self,
        system: &mut McSystem,
        state: McState
    ) -> Result<(), McError>;
    fn execution_mode(&self) -> &ExecutionMode;
    fn visited(&mut self) -> &mut VisitedStates;
    fn prune(&mut self) -> &mut PruneFn;
    fn goal(&mut self) -> &mut GoalFn;
    fn invariant(&mut self) -> &mut InvariantFn;
    fn collect(&mut self) -> &mut CollectFn;
    fn stats(&mut self) -> &mut McStats;

    // Provided methods
    fn process_event(
        &mut self,
        system: &mut McSystem,
        event_id: usize
    ) -> Result<(), McError> { ... }
    fn search_step(
        &mut self,
        system: &mut McSystem,
        event: EventOrId
    ) -> Result<(), McError> { ... }
    fn take_event(&self, system: &mut McSystem, event_id: usize) -> McEvent { ... }
    fn add_event(&self, system: &mut McSystem, event: McEvent) -> usize { ... }
    fn corrupt_message(&self, msg: Message) -> Message { ... }
    fn create_corrupted_receive(
        &self,
        event: McEvent,
        corrupted_msg: Message
    ) -> McEvent { ... }
    fn duplicate_event(&self, system: &mut McSystem, event_id: usize) -> McEvent { ... }
    fn debug_log(&self, event: &McEvent, depth: u64) { ... }
    fn have_visited(&mut self, state: &McState) -> bool { ... }
    fn mark_visited(&mut self, state: McState) { ... }
    fn on_final_state_reached(&mut self, status: String) { ... }
    fn check_state(&mut self, state: &McState) -> Option<Result<(), McError>> { ... }
    fn set_collect(&mut self, collect: CollectFn) { ... }

}

Trait with common functions for different model checking strategies.

Required Methods

fn build(config: StrategyConfig) -> Selfwhere Self: Sized,

Builds a new strategy instance.

fn run(&mut self, system: &mut McSystem) -> McResult

Launches the strategy execution.

fn search_step_impl( &mut self, system: &mut McSystem, state: McState ) -> Result<(), McError>

Callback which in called whenever a new system state is discovered.

fn execution_mode(&self) -> &ExecutionMode

Returns the used execution mode.

fn visited(&mut self) -> &mut VisitedStates

Returns the visited states set.

fn prune(&mut self) -> &mut PruneFn

Returns the prune function.

fn goal(&mut self) -> &mut GoalFn

Returns the goal function.

fn invariant(&mut self) -> &mut InvariantFn

Returns the invariant function.

fn collect(&mut self) -> &mut CollectFn

Returns the collect function.

fn stats(&mut self) -> &mut McStats

Returns the model checking execution stats.

Provided Methods

fn process_event( &mut self, system: &mut McSystem, event_id: usize ) -> Result<(), McError>

Explores the possible system states reachable from the current state after processing the given event. Calls search_step_impl for each new discovered state.

fn search_step( &mut self, system: &mut McSystem, event: EventOrId ) -> Result<(), McError>

Applies the specified event to the system, calls search_step_impl with the produced state and restores the system state afterwards.

fn take_event(&self, system: &mut McSystem, event_id: usize) -> McEvent

Takes the event from pending events by id.

fn add_event(&self, system: &mut McSystem, event: McEvent) -> usize

Adds the event to pending events list.

fn corrupt_message(&self, msg: Message) -> Message

Applies corruption to the Message.

fn create_corrupted_receive( &self, event: McEvent, corrupted_msg: Message ) -> McEvent

Creates MessageReceived event with corrupted msg.

fn duplicate_event(&self, system: &mut McSystem, event_id: usize) -> McEvent

Duplicates event from pending events list by id. The new event is left in pending events list and the old one is returned.

fn debug_log(&self, event: &McEvent, depth: u64)

Prints the log for particular event if the execution mode is Debug.

fn have_visited(&mut self, state: &McState) -> bool

Checks if the system state was visited before.

fn mark_visited(&mut self, state: McState)

Marks the system state as already visited.

fn on_final_state_reached(&mut self, status: String)

Adds new information to model checking execution statistics.

fn check_state(&mut self, state: &McState) -> Option<Result<(), McError>>

Applies user-defined checking functions to the system state and returns the result of the check.

fn set_collect(&mut self, collect: CollectFn)

Set collect predicate.